Fixes for EPOCH_CONTROL adaptive strategy

+++ profilers/src/globals.cpp
@@ -7,8 +7,8 @@
 /* Globals for sampling and adaptive profilers */
 uint64_t sp_initial_sample_size;
 uint64_t sp_sample_size;
-uint64_t sp_epoch_period;
-uint64_t sp_target_overhead;
+double sp_epoch_period;
+double sp_target_overhead;
 uint64_t g_total_overhead = 0; // Overhead incurred due to profiling
 uint64_t g_total_process_time = 0; // Total process time until last epoch sample
 uint64_t g_last_epoch_random = 0; // Random added to last epoch period
